# Factor out the routine for calculating efforts and improve speed by reducing printouts