Before, `typingEffort` calculation was in `carpalx`, but optimization requires calculating sum of typing efforts for different corpuses. Now the calculation is moved into `carpalx.ts`.

Also, logging in console take a considerable amount of time. Now the console logs only when there's an improvement or annealing so you can keep track of current score.

## Benchmarks

N = 100, 5 runs
Old average: 9.128s
New average: 8.534s (6.5% faster)

N = 1000, 2 runs
Old average: 82.535s
New average: 75.080s (9.0% faster)

-----
Note: I had to change display method because the refactoring moves logging into the `sumTypingEfforts` function. Once this function returns a value to the main loop, `console.clear()` got called, so it seems like nothing was printed. This end up saving some time ;)